bugs fixed

+++ qsm_algorithm/iLSQR/STISuite/parse_varargin_STISuiteiLSQR.m
@@ -19,35 +19,35 @@ params.padsize=[4,4,4];
 
 for kvar = 1:length(arg)
     if strcmpi(arg{kvar},'b0dir')
-        params.H = arg{kvar+1};
+        params.H = double(arg{kvar+1});
         continue
     end
     if strcmpi(arg{kvar},'tol_step1')
-        params.tol_step1 = arg{kvar+1};
+        params.tol_step1 = double(arg{kvar+1});
         continue
     end
     if strcmpi(arg{kvar},'tol_step2')
-        params.tol_step2 = arg{kvar+1};
+        params.tol_step2 = double(arg{kvar+1});
         continue
     end
     if strcmpi(arg{kvar},'iteration')
-        params.niter = arg{kvar+1};
+        params.niter = double(arg{kvar+1});
         continue
     end
     if strcmpi(arg{kvar},'TE')
-        params.TE = arg{kvar+1};
+        params.TE = double(arg{kvar+1});
         continue
     end
     if strcmpi(arg{kvar},'B0')
-        params.B0 = arg{kvar+1};
+        params.B0 = double(arg{kvar+1});
         continue
     end
     if strcmpi(arg{kvar},'threshold')
-        params.Kthreshold = arg{kvar+1};
+        params.Kthreshold = double(arg{kvar+1});
         continue
     end
     if strcmpi(arg{kvar},'padsize')
-        params.padsize = arg{kvar+1};
+        params.padsize = double(arg{kvar+1});
         continue
     end
 end
